+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Il formato utilizzato è basato sulle linee guida di [Keep a Changelog](http://k
 
 ### Aggiunto (Added)
  - Nuovo *Sconto finale* per gli **Ordini**, **Preventivi**, **DDT** e **Contratti**, influenza il valore *Netto a pagare* del documento.
+ - Nuovo filtro in attività per mostrare al tecnico solo le attività assegnate. 
 
 ## 2.4.22